+++ src/monagement/monagement_core/mod.rs
@@ -6,16 +6,16 @@ use crate::monagement::{level_core::FirstLevel, node_core::Node};
 
 #[derive(Debug)]
 pub struct MonagementCore {
-    pub(crate) max_size: u32,
-    pub(crate) bitmap: i32,
-    pub(crate) minimum_size: u32,
-    pub(crate) second_level_count: u32,
+    pub(crate) max_size: u64,
+    pub(crate) bitmap: u64,
+    pub(crate) minimum_size: u64,
+    pub(crate) second_level_count: u64,
     pub(crate) fl_list: Vec<FirstLevel>,
     pub(crate) linked_list: Vec<Option<Node>>,
     pub(crate) free_linked_list_index: Vec<usize>,
 
     //
-    update_counter: Option<(u32, u32, usize)>,
+    update_counter: Option<(u64, u64, usize)>,
     update_back_link: (Option<usize>, Option<usize>),
 }
 
@@ -32,7 +32,7 @@ impl MonagementCore {
         &self.fl_list
     }
 
-    pub fn bitmap(&self) -> i32 {
+    pub fn bitmap(&self) -> u64 {
         self.bitmap
     }
 }
